requirements.txt
bootstrap4==0.1.0
Django==3.2.5
django-allauth==0.45.0
django-bootstrap==0.2.4
django-ckeditor==6.1.0
django-chartjs==2.2.1
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-mysql==3.12.0
django-tinymce4-lite==1.8.0
djangorestframework==3.12.4
django-widget-tweaks==1.4.8
gunicorn==20.1.0
jmespath==0.10.0
mysql-client==0.0.1
mysql-connector==2.2.9
Pillow==8.3.1
pylint==2.9.5
python-dateutil==2.8.2
soupsieve==2.2.1
whitenoise==5.3.0
